# back-end/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de manipulação de do sistema",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/",
      "description": "Api de teste"
    }
  ],
  "paths": {
    "/carros": {
      "post": {
        "summary": "Cadastro de carros",
        "tags": ["Carros"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Carro"
              },
              "examples": {
                "carro": {
                  "value": {
                    "nomeVeiculo": "Toro",
                    "placa": "MOJ-6356",
                    "combustivel": "Gasolina",
                    "cor": "Cinza",
                    "ano": "2019",
                    "valor": "150632"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "400": {
            "description": "Car already exists"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Carro"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Listagem de todos carros",
        "tags": ["Carros"],
        "responses": {
          "200": {
            "description": "Lista de carros",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Carro"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/carros/{id}": {
      "put": {
        "summary": "Atualiza um carro existente pelo ID",
        "description": "Atualiza um carro existente pelo ID informado no parâmetro",
        "tags": ["Carros"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": "true",
            "schema": {
              "type": "string"
            },
            "description": "ID do carro a ser atualizado"
          },
          {
            "in": "body",
            "name": "carro",
            "required": "true",
            "schema": {
              "type": "object",
              "properties": {
                "nomeVeiculo": {
                  "type": "string",
                  "description": "Nome do veiculo"
                },
                "placa": {
                  "type": "string",
                  "description": "Placa do veiculo"
                },
                "combustivel": {
                  "type": "string",
                  "description": "Tipo de combustivel do veiculo"
                },
                "cor": {
                  "type": "string",
                  "description": "Cor do veiculo"
                },
                "ano": {
                  "type": "number",
                  "description": "Ano do veiculo"
                },
                "valor": {
                  "type": "number",
                  "description": "Valor do veiculo"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Carro atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Carro"
                }
              }
            }
          },
          "400": {
            "description": "Nao foi possivel atualizar o carro"
          }
        }
      },
      "delete": {
        "summary": "Remover um carro existente pelo ID",
        "description": "Remover um carro existente pelo ID informado no parâmetro",
        "tags": ["Carros"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": "true",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Carro removido com sucesso"
          },
          "400": {
            "description": "Nao foi possivel remover o carro"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login do usuario",
        "description": "Efetua o login do usuario no sistema",
        "tags": ["Usuario"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "senha": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login realizado com sucesso"
          },
          "400": {
            "description": "Senha invalida"
          },
          "404": {
            "description": "Usuario não encontrado"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": ["Usuario"],
        "summary": "Cadastrar usuário",
        "description": "Função para cadastro de usuario",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "senha": {
                  "example": "any"
                },
                "imagem": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cadastrado com sucesso"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Usuario já existente"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "tags": ["Usuario"],
        "summary": "Recuperar senha",
        "description": "Função para recuperar a senha do usuário",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "tags": ["Usuario"],
        "summary": "Resetar a senha",
        "description": "Função para resetar a senha do usuário",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "token": {
                  "example": "any"
                },
                "novaSenha": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Carro": {
        "type": "object",
        "properties": {
          "nomeVeiculo": {
            "type": "string"
          },
          "placa": {
            "type": "string"
          },
          "combustivel": {
            "type": "string"
          },
          "cor": {
            "type": "string"
          },
          "ano": {
            "type": "number"
          },
          "valor": {
            "type": "number"
          }
        }
      },
      "Usuario": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "senha": {
            "type": "string"
          },
          "tokenResetarSenha": {
            "type": "string"
          },
          "tokenResetarSenhaExpirado": {
            "type": "date"
          },
          "dataCriacao": {
            "type": "date"
          },
          "imagem": {
            "type": "string"
          },
          "administrador": {
            "type": "boolean"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}